Generalize TransformFeedbackSimpleTests for device_address_commands
Generalize some of tests in vktTransformFeedbackSimpleTests.cpp to cover:
vkCmdBindTransformFeedbackBuffers2EXT,
vkCmdBeginTransformFeedback2EXT,
vkCmdEndTransformFeedback2EXT and
vkCmdDrawIndirectByteCount2EXT.

Components: Vulkan

VK-GL-CTS issue: 5498

Affects:
dEQP-VK.transform_feedback.*

+++ external/vulkancts/modules/vulkan/draw/vktDrawInstancedTests.cpp
@@ -1175,7 +1175,7 @@ InstancedTests::InstancedTests(tcu::TestContext &testCtx, const SharedGroupParam
         {
             // reduce number of tests for dynamic rendering cases where secondary command buffer is used
             if (groupParams->useSecondaryCmdBuffer &&
-                ((topology % 2u) == groupParams->secondaryCmdBufferCompletelyContainsDynamicRenderpass))
+                ((topology % 2u) == uint32_t(groupParams->secondaryCmdBufferCompletelyContainsDynamicRenderpass)))
                 continue;
 
             for (auto function : functions)
